// src/app/interfaces/pokemon.ts
export interface CustomPokemon{
    name: string,
    id: number,
    types: string[],
    image?: string
}